Move the files under kml/modules instead
Moved
* local_global_conversion.py
* drone_odometry_local.py
from mavlink/modules to kml/modules
