+++ kml/modules/drone_odometry_local.py
@@ -2,7 +2,7 @@
 Drone odometry in local space (origin at home location).
 """
 
-from . import drone_odometry
+from ...mavlink.modules import drone_odometry
 
 
 class DronePositionLocal:
